+++ NEWS.md
@@ -5,8 +5,14 @@
 * Added `save_hashmap` and `load_hashmap` function for saving and loading 
   `Hashmap` objects (#3).
   
-* Added S3 `plot` method
+* Added S3 `plot` method.
+
+* Added `clone` function for creating deep copies. This is also reflected in 
+  the C++ API via `HashMap::clone`. 
 
 ## Improvements
 
-* Integers are now printed correctly
+* Integers are now printed correctly. 
+
+* Key and value vector access is now automatically cached on the first call to 
+  `$keys()` and `$values()`, respectively.